+++ sw/tests/src/fcvt_d_wu_copift.c
@@ -16,7 +16,7 @@ int main() {
 
     // Allocate input and output arrays
     uint64_t *input = (uint64_t *)snrt_l1_alloc_compute_core_local(
-        LENGTH * sizeof(uint32_t), sizeof(uint32_t));
+        LENGTH * sizeof(uint64_t), sizeof(uint64_t));
     double *golden_output = (double *)snrt_l1_alloc_compute_core_local(
         LENGTH * sizeof(double), sizeof(double));
     double *actual_output = (double *)snrt_l1_alloc_compute_core_local(
